DataGeneration/requirements.txt
matplotlib
numpytensorflow 
kagglehub